% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_daily}
\alias{convert_to_daily}
\title{Convert interval data into daily data}
\usage{
convert_to_daily(Data, kWh = TRUE, thresh = 24)
}
\arguments{
\item{Data}{A dataframe that contains interval data}

\item{kWh}{A logical if TRUE it means that the interval data are energy data}
}
\value{
A dataframe with hourly interval data
}
\description{
\code{convert_to_day} This function convert interval data into daily data
}
