% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nre.R
\name{dissim_calc}
\alias{dissim_calc}
\title{Calculate dissimilarity metric}
\usage{
dissim_calc(post_Data, predicted_post_Data)
}
\arguments{
\item{post_Data}{A dataframe that contains the actual post energy use time series}

\item{predicted_post_Data}{A dataframe that contains the predicted post energy use time series}
}
\value{
A one column matrix where the value in each line correspond 
the dissimilarity metric of the corresponding day.
}
\description{
\code{dissim_calc} This function calculate CORT dissimilarity metric for 
each day of the energy use time series
}
