% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nre.R
\name{nre_detect}
\alias{nre_detect}
\title{Detect potential Non-Routine Events}
\usage{
nre_detect(post_Data, predicted_post_Data)
}
\arguments{
\item{post_Data}{A dataframe that contains the actual post energy use time series}

\item{predicted_post_Data}{A dataframe that contains the predicted post energy use time series}
}
\value{
a nre_detect object, which is a list with the following components:
\describe{
  \item{dissim_tab}{A dataframe that correspond to the computed daily dissimilarities}
  \item{cpt_res}{An object of S4 class "cpt" (refer to changepoint package 
documentation for more information)}
  \item{nre_dates}{A vector containing character string that corresponds to 
potential non-routine events dates}
}
}
\description{
\code{nre_detect} This function uses a statistical change point algorithm 
(i.e., \code{cpt.meanvar} function from "changepoint" package) to detect 
potential non-routine events
}
