% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nre.R
\name{nre_idx}
\alias{nre_idx}
\title{Indices of non-routine events}
\usage{
nre_idx(Data, nre_dates)
}
\arguments{
\item{Data}{A dataframe that contains post period energy use time series data}

\item{nre_dates}{A dataframe that contains the start and the ends of each non-routine
period event}
}
\value{
A vector of indices of observations that correspond 
to the non-routine events
}
\description{
\code{nre_idx} This function extract the indices of observations that correspond 
to the non-routine events defined by the user through nre_dates.
}
